Implement `confgen set path key` functionality
Updating inventory with this command will add necessary keys to existing
config.yaml files or it will create new ones.
The way of dumping structs to yaml files in transparent to the library,
but the representation changes. Because while setting single key, the
entire inventory is rebuilt (simpler) first build might seem weird.

Notation <key>: <value> will be changed to {<key>: <value>}
-mysql: 2.0
+{mysql: 2.0}

Also strings are represented differently:

!!python/unicode 'secret': !!python/unicode 'sUpEr'}

I am not sure if this is a implementation bug or if it's a valid yaml
representation of python structures.